+++ event-handlers/src/test/java/handlers/UpsertNviCandidateHandlerTest.java
@@ -22,8 +22,8 @@ import com.amazonaws.services.lambda.runtime.events.SQSEvent.SQSMessage;
 import java.util.List;
 import java.util.UUID;
 import java.util.stream.Stream;
-import no.sikt.nva.nvi.common.model.dao.Candidate;
-import no.sikt.nva.nvi.common.model.dao.Level;
+import no.sikt.nva.nvi.common.model.business.Candidate;
+import no.sikt.nva.nvi.common.model.business.Level;
 import no.sikt.nva.nvi.common.model.events.CandidateEvaluatedMessage;
 import no.sikt.nva.nvi.common.model.events.CandidateStatus;
 import no.sikt.nva.nvi.common.model.events.NviCandidate.CandidateDetails;
