+++ nvi-commons/src/main/java/no/sikt/nva/nvi/common/model/business/Level.java
@@ -1,23 +1,24 @@
-package no.sikt.nva.nvi.common.model.dao;
+package no.sikt.nva.nvi.common.model.business;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
 import java.util.Arrays;
 
-public enum CompletionStatus {
-    COMPLETED("Completed"), IN_PROGRESS("In progress");
+public enum Level {
+    UNASSIGNED("Unassigned"), LEVEL_ONE("1"), LEVEL_TWO("2");
 
     @JsonValue
     private final String value;
 
-    CompletionStatus(String value) {
+    Level(String value) {
+
         this.value = value;
     }
 
     @JsonCreator
-    public static CompletionStatus parse(String value) {
+    public static Level parse(String value) {
         return Arrays
-                   .stream(CompletionStatus.values())
+                   .stream(Level.values())
                    .filter(level -> level.getValue().equalsIgnoreCase(value))
                    .findFirst()
                    .orElseThrow();
