+++ nvi-commons/src/main/java/no/sikt/nva/nvi/common/model/business/PublicationDate.java
@@ -1,4 +1,4 @@
-package no.sikt.nva.nvi.common.model.dao;
+package no.sikt.nva.nvi.common.model.business;
 
 public record PublicationDate(String year, String month, String day) {
