+++ nvi-commons/src/main/java/no/sikt/nva/nvi/common/model/business/Status.java
@@ -1,25 +1,24 @@
-package no.sikt.nva.nvi.common.model.dao;
+package no.sikt.nva.nvi.common.model.business;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
 import java.util.Arrays;
 
-public enum Level {
-    UNASSIGNED("Unassigned"), LEVEL_ONE("1"), LEVEL_TWO("2");
+public enum Status {
+    APPROVED("Approved"), PENDING("Pending"), REJECTED("Rejected");
 
     @JsonValue
     private final String value;
 
-    Level(String value) {
-
+    Status(String value) {
         this.value = value;
     }
 
     @JsonCreator
-    public static Level parse(String value) {
+    public static Status parse(String value) {
         return Arrays
-                   .stream(Level.values())
-                   .filter(level -> level.getValue().equalsIgnoreCase(value))
+                   .stream(Status.values())
+                   .filter(status -> status.getValue().equalsIgnoreCase(value))
                    .findFirst()
                    .orElseThrow();
     }
