+++ nvi-commons/src/main/java/no/sikt/nva/nvi/common/model/business/Username.java
@@ -1,0 +1,5 @@
+package no.sikt.nva.nvi.common.model.business;
+
+public record Username(String value) {
+
+}
