+++ nvi-commons/src/main/java/no/sikt/nva/nvi/common/model/business/VerifiedCreator.java
@@ -1,4 +1,4 @@
-package no.sikt.nva.nvi.common.model.dao;
+package no.sikt.nva.nvi.common.model.business;
 
 import java.net.URI;
 import java.util.List;
