--- nvi-commons/src/main/java/no/sikt/nva/nvi/common/model/dao/Username.java
@@ -1,5 +1,0 @@
-package no.sikt.nva.nvi.common.model.dao;
-
-public record Username(String value) {
-
-}
