+++ nvi-commons/src/main/java/no/sikt/nva/nvi/common/service/NviService.java
@@ -4,12 +4,12 @@ import static no.sikt.nva.nvi.common.model.events.CandidateStatus.CANDIDATE;
 import java.net.URI;
 import java.util.List;
 import no.sikt.nva.nvi.common.NviCandidateRepository;
-import no.sikt.nva.nvi.common.model.dao.ApprovalStatus;
-import no.sikt.nva.nvi.common.model.dao.Candidate;
-import no.sikt.nva.nvi.common.model.dao.Level;
-import no.sikt.nva.nvi.common.model.dao.PublicationDate;
-import no.sikt.nva.nvi.common.model.dao.Status;
-import no.sikt.nva.nvi.common.model.dao.VerifiedCreator;
+import no.sikt.nva.nvi.common.model.business.ApprovalStatus;
+import no.sikt.nva.nvi.common.model.business.Candidate;
+import no.sikt.nva.nvi.common.model.business.Level;
+import no.sikt.nva.nvi.common.model.business.PublicationDate;
+import no.sikt.nva.nvi.common.model.business.Status;
+import no.sikt.nva.nvi.common.model.business.VerifiedCreator;
 import no.sikt.nva.nvi.common.model.events.CandidateEvaluatedMessage;
 import no.sikt.nva.nvi.common.model.events.NviCandidate.CandidateDetails;
 import no.sikt.nva.nvi.common.model.events.NviCandidate.CandidateDetails.Creator;
