+++ nvi-commons/src/test/java/no/sikt/nva/nvi/common/CandidateTest.java
@@ -15,15 +15,15 @@ import java.time.LocalDate;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
-import no.sikt.nva.nvi.common.model.dao.ApprovalStatus;
-import no.sikt.nva.nvi.common.model.dao.Candidate;
-import no.sikt.nva.nvi.common.model.dao.Level;
-import no.sikt.nva.nvi.common.model.dao.Note;
-import no.sikt.nva.nvi.common.model.dao.Period;
-import no.sikt.nva.nvi.common.model.dao.PublicationDate;
-import no.sikt.nva.nvi.common.model.dao.Status;
-import no.sikt.nva.nvi.common.model.dao.Username;
-import no.sikt.nva.nvi.common.model.dao.VerifiedCreator;
+import no.sikt.nva.nvi.common.model.business.ApprovalStatus;
+import no.sikt.nva.nvi.common.model.business.Candidate;
+import no.sikt.nva.nvi.common.model.business.Level;
+import no.sikt.nva.nvi.common.model.business.Note;
+import no.sikt.nva.nvi.common.model.business.Period;
+import no.sikt.nva.nvi.common.model.business.PublicationDate;
+import no.sikt.nva.nvi.common.model.business.Status;
+import no.sikt.nva.nvi.common.model.business.Username;
+import no.sikt.nva.nvi.common.model.business.VerifiedCreator;
 import no.unit.nva.commons.json.JsonUtils;
 import org.junit.jupiter.api.Test;
