+++ nvi-commons/src/test/java/no/sikt/nva/nvi/common/service/NviServiceTest.java
@@ -15,8 +15,8 @@ import static org.hamcrest.Matchers.is;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
-import no.sikt.nva.nvi.common.model.dao.Candidate;
-import no.sikt.nva.nvi.common.model.dao.Level;
+import no.sikt.nva.nvi.common.model.business.Candidate;
+import no.sikt.nva.nvi.common.model.business.Level;
 import no.sikt.nva.nvi.common.model.events.CandidateEvaluatedMessage;
 import no.sikt.nva.nvi.common.model.events.CandidateStatus;
 import no.sikt.nva.nvi.common.model.events.NviCandidate.CandidateDetails;
