+++ nvi-test/src/main/java/no/sikt/nva/nvi/test/TestUtils.java
@@ -7,10 +7,10 @@ import java.util.List;
 import java.util.Random;
 import java.util.UUID;
 import java.util.stream.Stream;
-import no.sikt.nva.nvi.common.model.dao.ApprovalStatus;
-import no.sikt.nva.nvi.common.model.dao.PublicationDate;
-import no.sikt.nva.nvi.common.model.dao.Status;
-import no.sikt.nva.nvi.common.model.dao.VerifiedCreator;
+import no.sikt.nva.nvi.common.model.business.ApprovalStatus;
+import no.sikt.nva.nvi.common.model.business.PublicationDate;
+import no.sikt.nva.nvi.common.model.business.Status;
+import no.sikt.nva.nvi.common.model.business.VerifiedCreator;
 import no.sikt.nva.nvi.common.model.events.NviCandidate.CandidateDetails;
 import no.sikt.nva.nvi.common.model.events.NviCandidate.CandidateDetails.Creator;
 import nva.commons.core.paths.UriWrapper;
